Exporter now uses non-blocking sends
This means we have to deal with buffering partial sends, so the
buffering code has become a bit more complicated.

Fixed bad buffer shrinking code which would crash whenever used.

Fixed leaking of addrinfo structures when attempting to connect
to a target.

Connection failures when sending buffered data to a target are
now caught and the connection is closed, rather than continually
failing to send.

Use level-triggered epoll events rather than edge triggered ones
in the exporter. This fixes two problems: the full queue deadlock
where the exporter wouldn't read from the queue because the
processing thread couldn't write to it (because it was full),
and the missed timer bug which would cause the exporter main
function to loop forever.