Add new "no more mediators" message to communication protocol.
This message will be used by the provisioner to announce to a
collector that it has sent details about all of the mediators to
the collector.

Upon receipt, the collector will use this message to purge any
mediators that it had in memory that appear to be no longer active.

This is required to handle the scenario where the provisioner
restarts *and* the mediator set changes at the same time.