Start adding SIP processing to the collector
Collector now recognises SIP packets which have URIs that are
known as being part of active intercepts. The collector processing
threads will forward these packets to the sync thread for
further processing.

The sync thread takes the SIP packets, matches them to the correct
LIID(s) and attempts to match the call to an existing CIN. If none
exists, a CIN is created by hashing the Call Id. If any RTP stream
info is present, this is extracted from the SDP body.

Next step: push the RTP stream info back to the collector so it
can hopefully intercept the RTP traffic as well.