Mediator config can now be reloaded via a SIGHUP.
If the provisioner details change, the mediator will connect to
the new provisioner.

If the listening socket details change, the mediator will
announce itself anew to the provisioner. The provisioner will
recognise that the mediator details have changed and will first
tell each collector to withdraw the old mediator details, then
send the new details to all of the collectors.

On the collector side, these withdrawal and annoucement messages
bounce through the sync thread to the export thread. The export
thread acts upon them by dropping the old mediator connection and
connecting to the newly announced one.