Add ability to change inputs by reloading collector config.
This includes changing the number of threads, although the input
will be destroyed and re-created to do this.

Inputs may be added, deleted or modified.

"Removed" inputs are placed in a list for safe-keeping (we can't
destroy the traces while there are outstanding packets, e.g. at
the exporter). They are freed when the collector exits.