Update collector to use new libwandder encoding API
Note that the exporter does not release results back to the
encoder that created them -- thread safety issues mean that this
requires more work to be able to do correctly -- so we are still
being wasteful with the encoded result structures.

However, performance during the main packet processing loop should
still improve due to recycling of the internal wandder single
data item structures.