Replace non-blocking queue reads with blocking ones
The reasoning behind this is that the non-blocking read ends up
holding the lock for the duration of the read(), whereas the
blocking read does not. Since we're only going to read from the
queue when epoll signals the queue fd as ready for reading, we
can trust that there'll be something in the queue 99% of the time
that we try to read from it, so we can afford to "block".

The impact of this is that the threads that are trying to put
items in the queue don't end up spinning on the lock whenever the
queue is being read from. This allows the threads to quickly push
something on the queue and then get back to their main job of
reading packets.