Fixed unwanted collector->mediator disconnects
When the provisioner restarted, collectors would always disconnect
from their mediators even when the mediators were unchanged.

The problem was due to the provisioner deciding to send the
"no more mediators" message before the mediators had had a chance
to reconnect to it.

New approach is to only disconnect mediators that have not been
announced within 10 seconds of the last mediator announcement
since the provisioner disappeared.

This still has some downsides:
  * the timer doesn't start until the provisioner announces at
    least one mediator, so if there are no legit mediators, then
    you will keep exporting to the old ones (maybe not a bad
    thing?).
  * in theory, the no-longer-valid mediators could be made to
    persist forever if you kept connecting new mediators to the
    provisioner at < 10 second intervals. Again, probably not a
    genuine problem -- you shouldn't have rogue mediators able to
    connect to your provisioner and there is probably no harm in
    an old mediator continuing to receive records anyway.