Fix double free issues with user intercept lists.
We needed a distinct hash handle for the user intercept list even
though the user intercept list and the IP intercept list use the
same key.

Otherwise the hash table maintenance gets very confused and
transfers "next" links from one list to the other, so we end
up iterating over intercepts that shouldn't be in a particular
list to begin with.