Add ability to configure "known" RADIUS servers
These servers will be used by the collectors to recognise RADIUS
traffic and parse it accordingly.

The implementation is built on the concept of "core" servers, so
we can use the same messages and data structures to implement
similar functionality for other key protocols which need to be
forwarded to the sync thread, e.g. SIP and DHCP.

The configuration of known servers happens at the provisioner and
this config is then passed down to all active collectors. As with
other config, the implementation is designed to handle config
reloads and temporary provisioner disappearances.

NOTE: I am aware that RADIUS traffic typically uses different
ports for auth and accounting traffic. The system will still work
in this case: you'll just have to have two entries in the config
for each RADIUS server. Alternatively, if you're using the default
ports then you can just specify the IP only and OpenLI will handle
the rest (code for this is still to come...).