Implemented outline for first session tracking plugin (RADIUS)
None of the plugin callbacks do anything yet, but this is what
I'm planning a AAA parsing plugin to initially look like.

Re-worked free_single_session() to take the session itself as
a parameter, rather than having to look it up. In all cases so
far, we already have a pointer to the session when we want to
free it so don't bother looking it up again.