Reverted previous requirement of a user session for ALU intercepts
The session ID in the ALU shim header can be used as our CIN so
there's no requirement for the sync thread to see the RADIUS
session for the user to re-encode the ALU packets properly.

In other words, announce any ALU intercepts to the collector
threads immediately.

This has the added benefit that the RADIUS IRIs can be generated
by a completely separate collector instance, which might be handy.