Rename "requestaction" and "responseaction"
These names were misleading -- if a packet was a correctly
ordered response, then 'requestaction' was being updated.
Instead, I've named them 'firstaction' and 'secondaction' to
reflect the fact that there are up to two actions that can
be applied per captured RADIUS packet.

Also return the first non-NONE action to the caller, just
in case they need it. We will be handling both actions
properly inside our create_iri function, but an indication
of some action needing to take place might be useful.