Combine "access request" and "accounting request" into one struct
There's no real need for this distinction in practice and this
will greatly simplify our code.

Also added a saved attribute list to the request struct as well,
since we'll need them for IRIs that are generated on the response
but need access to attributes from the request. Not populated yet,
but it's there.