Avoid constant malloc and free of att_vals
Instead, everytime we create an attribute: malloc some space
for att_val and don't free it when we release the attribute;
reuse that space next time we recycle the attribute.

Downside: we will always "copy" every attribute, even for packets
where we would have previously avoided it (i.e. non-orphan
responses). This could be fixed by having two freelists: one
for attributes with malloced memory and one for attributes
that point to bytes in the packet itself.

Advantage: saving the attributes when we have to store them for
a later packet is much easier -- we already have a list full of
copies so just move it from the parsed packet to the saved data.