Assign "unique" CINs to RADIUS sessions
The CIN is derived from the Acct-Session-ID, if present. If not,
we'll use the Authenticator bytes, provided the packet is an
Access-Request. If all else fails, use the timestamp and hope that
we don't get any clashes.

Realistically, we should never fall through to the timestamp
option but it's better than just throwing a tantrum.