Add etsili_generic_t type + APIs to etsili_core
For IPIRI contents, there are a lot of optional fields that
may or may not be available when constructing the message. Rather
than having 24 potentially mostly empty parameters to populate
and pass in at IRI generation time, we can just provide a list
of whatever we've got available and have the IRI construction
code deal with that.