Initial code for encoding IPIRIs
Not all IPIRI fields are supported yet -- just the ones that
match standard data types.

Code is not actually invoked yet, still need to write the code
in the RADIUS plugin to extract the necessary fields and then
call the new ip_iri() function for each type of IRI message.