Mediator keep-alives contain more useful information now.
Specifically, the user can configure a network operator ID that
will be included in the keep alive. The mediator ID is also
included as the "network element ID" so the agencies should be
able to tell who is (and isn't) sending them keep alives.

The LIID for a keep alive is a static string that identifies it
as being generated by OpenLI.