Keepalive response requirements can now change on the fly.
If the KA response config for an agency changes and the
provisioner re-announces that agency to the mediator, the
mediator config for that agency will change accordingly
(without dropping the handover connections).

Code is a bit nasty looking, but it seems to work at least.

Also fixed a bug where the IP addresses and ports for
re-announced handovers were being freed prematurely.