Add ability to configure user access type in provisioner config
The "access type" is an essential field in IPIRIs and is not
obviously able to determined programatically. Instead, I've
added an extra config option ('accesstype') to the ipintercept
config which will allow the operator to manually set this
field to the right value.

I figure the operator themselves will be best placed to know
what they should put in here when they receive the warrant, since
they (hopefully) know what type of service they are providing to
the target.