Don't require both leading and trailing quotes for stripping.
Stripping the trailing quote is permanent, but we often will end
up trying to strip the same username or realm multiple times.
Therefore, we need to accept situations where there is only a
leading quote and make sure we strip that.

Otherwise we'll end up with plenty of failed SIP user / realm
matches because the terms came out of our parser as
"username and "realm.