Add ability to write intercepts to pcap files.
Live streaming to an agency is the ETSI way, but there may be some
situations where you are dealing with a non-ETSI compliant agency
and they just want some old fashioned pcaps.

To use pcap mode, set the agencyid for the intercept to
"pcapdisk". Also ensure that the "pcapdirectory" config option is
set in the mediator config, so that it knows where to write the
files.