Partial implementation of moving ETSI encoding to the exporter(s)
Very much incomplete -- this is mostly just a backup via commit.

The goal is to do all of our ETSI encoding in exporter threads,
so we can avoid issues with different parts of communication
sessions ending up on different processing threads.

Instead, we will now send "jobs" to the exporting threads, where
the dest thread is determined by hashing the LIID so all
packets for the same intercept end up being encoded in the same
place.

This is a big rewrite of the collector, so there is a lot more
still to do but I think I've got all of the sync and processing
threads to the point where they are generating the new job
messages and will call into the new export queue manager API to
ensure messages get pushed to the right thread(s).

Still to do:
 * write the code for the export queue manager API
 * update the exporter code to process the new messages
 * actually do the ETSI encoding (mostly just a case of tweaking
   the old encoding code to take a 'job' as input).