Add exporter code for handling INTERCEPT_DETAILS and INTERCEPT_OVER
Changed how these messages are formed and sent to avoid issues
with multiple references to the same string (which might then
be freed before everyone is done with it). The details get
duplicated for every exporter thread, which may seem a bit
wasteful, but hopefully we're not doing this too often such that
it matters.

The alternative is reference counts and mutexes, so we can try
that if the duplication + memory allocs + frees is proving to
be problematic.

Also, each exporter thread no longer has a reference to the
RADIUS plugin -- instead, a pointer to the sync thread's reference
is included in the job message, which should be enough to make
things work properly.