Improved thread-safety and stopped sending PACKET_FINs
PACKET_FINs were proving problematic as they had an annoying
tendency to be received *before* the encoding jobs that used the
packet, causing the packet to be freed and then used (i.e.
segfaulting).

Now, we simply increment the ref counter before pushing a job onto
the queue and the exporter threads decrement the ref counter when
they finish the job.

Other threading issues fixed while sorting this out:
 * Better locking when touching the collector queues within
   the sync threads (especially during shutdown).
 * Block all signals before creating sync and export threads.
 * Lock before destroying the exporter epoll events, just in
   case someone is still registering new queues at the same time.
 * Lock the collector queues when pushing core server lists, just
   in case a new collector registers with the sync thread at the
   same time.
 * Better logging as to why a SIP update failed.
 * Double check that there is a message available before reading
   from a collector->voipsync queue.