Added timer to trigger flushing and rotating of pcap output files.
This helps resolve the issues where the pcap files are incomplete
unless the mediator is halted (causing the output file to be
closed and flushed properly).

The flush alone is not perfect -- trailers etc are missing so the
flushed file is still incomplete -- but it will at least mean that
the file size will grow over time rather than appearing empty if
there is little intercepted activity.

File rotation will close the file properly and start a new one, so
rotated files are readable. The rotated files are now labelled with
a timestamp, so each will have a distinct filename. The files will
need to be merged (e.g. using tracemerge) to create a single pcap
for the whole intercept.

The flush support does require a cutting-edge libtrace and wandio,
which may be a bit awkward right now with the varying states of
test deployments. Since most users don't *require* the pcap output
feature, it may be worth making this an optional feature at build
time so only the people who really want this feature have to deal
with updating a whole bunch of dependencies...