Add mechanism for specifying static IP ranges as intercept targets
Everything so far just ensures that the ranges propagate into
the collector threads properly, where they are inserted into a
patricia tree. Each tree node stores a list of LIIDs that have
been tagged with that prefix.

The trees are completely unused when it comes to recognising
packets that should be intercepted -- that's the next phase of
development...