Add ability to set the CIN for a static IP range
Normally, we'd derive the CIN from the session ID for the AAA
session but a static assignment doesn't really have any AAA for
us to observe.

Instead, we can at least allow the operator to choose an
appropriate CIN to tag each static range with (or use the same
CIN for all ranges -- that's fine too).

If no session ID/CIN is provided, we just use 1 as the CIN.