Add ability to enable TCP keep-alives for handover sockets.
Despite being told to not use this in the ETSI spec, there are
always going to be firewalls etc that will trash TCP sessions
that are not responding to keep-alives.

All other internal sockets in OpenLI still don't use TCP
keep-alives but can be easily switched over to do so if
required.