Add IP fragment reassembly to SIP parsing code
Have seen some instances where UDP SIP Invites are so large that
they get fragmented. This code aims to reassemble any IP layer
fragments into coherent frames that can be properly parsed by osip.

Totally untested as yet with fragmented traffic, but seems to not
break horribly for the 99.9% of traffic that is not fragmented.

In theory, we will also be able to handle fragmented TCP SIP as
well, although I really hope that is something I never have to
worry about in practice.