Deal with padding in small SIP frames.
A small SIP message (e.g. a keep-alive) may be padded to
ensure that it has a minimum 64 bytes wire length, so we
cannot rely solely on 'remaining' to determine how much SIP
content is available in a packet. Instead, we use
min(remaining, payloadlen - sizeof(transhdr)) as our SIP
content length.