Fix potential race condition with attribute freelist
I have observed a situation where the attribute linked list
for a parsed RADIUS packet ends up containing a loop and suspect
that this was likely caused by a race between an export thread
releasing attributes at the same time as the RADIUS thread was
peeling some from the freelist.

Added a mutex that should prevent both actions happening
concurrently. The mutex is non-blocking, so the freelist won't
be used if either thread is unable to obtain the lock right away.
The additional of a mutex will probably impact on our performance
a bit, but it is for RADIUS traffic only so we can probably afford
the hit.

Since the bug is a rare occurence, I'm still testing to confirm
that it is resolved by this change.