Add ability to detect silent IP session logoff
Silent logoff occurs when an IP is assigned to a new user without
any obvious indication that the previous user's session had ended
(i.e. the RADIUS packets for the session end were lost before
our monitor). If we see a silent logoff for an intercept target,
we need to halt the interception immediately.

Fixed a variety of problems with the RADIUS plugin that were
causing us to miss actions or try to use the wrong attributes
to generate IRIs (especially when dealing with orphaned responses).

Created a structure specifically for storing an IP (range) that is
associated with a user session, rather than repeatedly adding the
same three fields to each structure that needs this information.
This also allows us to use the IP as a key in a uthash map.

Removed unnecessary PACKET_FIN messages from collector_sync.

Fixed bug where the IRI sequence number would be incremented,
even if no IRI was generated. Also fixed the erroneous error
reports about an "invalid iteration" when generating RADIUS IRIs,
which occured when secondaction was set but firstaction was not.