Outgoing socket connection attempts now time out properly.
Previously, we were doing blocking connect()s, which would
not time out for a very long time and effectively hang whatever
component was trying to make the connection.

Now connect()s are performed as non-blocking and a select() is
used to determine whether the connect has succeeded or timed
out.