Move agency connection attempts into a separate thread.
This should resolve issues where the mediator was blocking
regularly, waiting for a connection attempt to timeout rather
than concentrating on exporting messages to the agencies that we
are connected to.

Also fixed race where some of the pcap thread data was being
destroyed before the pcap thread had exited cleanly.