Fix continuous memory allocation in netcomms code.
The receive buffer for netcomms messages would continually grow
in size, even when there was plenty of reusable space in the front
of the buffer that we could re-use instead.

The code was smart enough to re-use that space, but would go ahead
with the buffer size increase regardless. This meant the mediator
would continue to use more and more memory over time, depending
on the amount of intercepted traffic that was being forwarded to
it.