Replace coreserver_match() with macro functions
We were calling coreserver_match() a lot for each packet, so we
were wasting a lot of time in function call overhead.

Moved populate_addrinfo into util.c so that it can be accessed
by the code that used to call coreserver_match().

Introduced a brief delay (temporarily) between starting the
exporter threads and the sync threads, just to ensure the exporters
will be ready to receive key info from the sync threads as soon
as they start up.