Partial implementation of new encoder-worker architecture
The main goal of the new architecture is to replace the exporter
threads with a single exporter thread (which manages per-intercept
state, seqnos etc as well as sending encoded records to the
mediator) plus a set of encoder worker threads (which encode
the ETSI records).

This should allow us to make better use of available CPUs for
encoding, which is the expensive part of the exporter.

This is a pretty big change and isn't complete yet -- this commit
mostly deals with getting encoding workers up and running, but the
encoded results are ignored by the exporter thread right now.