Phase 1 of yet another exporter re-structure
Incomplete, but the main idea is to split the old exporter into
two components: sequence tracker and forwarder.

Sequence trackers does all the sequence number tracking for CINs
and LIIDs, and turns intercepted packets into encoding jobs for
the encoding workers.

Forwarders take the encoded records from the encoding workers and
send them on to the appropriate mediator. They manage the
connections to the mediators, so need to receive all of the
mediator-related updates from the sync thread as well.

The main purpose of this rewrite to fix potential issues where
the exporting was getting deadlocked due to full queues, as well
as tidy up the whole exporting code in general. With any luck,
the clear separation of work into simple FIFO-style pipes will
also help us diagnose performance issues more quickly too and
maybe we will finally be able to get on top of the problems I've
been battling over the past month or two.