Further progress towards the exporter rewrite
This commit updates the forwarder to be able to add new mediators
and connect to them.

It also receives (and then immediately releases) encoded results
from the encoders.

Both encoded results and their published jobs are now released
back to their original allocators for re-use rather than being
freed. This is done in batches to try and improve performance.

Attempted to fix issues with the job freelist mutexes being
destroyed before other threads had finished, causing problems
when those other threads try to release jobs that they've
finished with.

Sync threads will now attempt to send a halt message to other
threads no more than 10 times before assuming that the other pull
socket is inactive -- this prevents deadlocks on exit where the
sync thread is blocking trying to send a halt to a thread that
has already halted due to an error.