More performance enhancements for high packet rates.
Removed the freelists for published jobs and encoded
records, as these were actually making performance worse.
The locking etc. required to safely update or use any of the
freelists appeared to be the main problem.

Instead, we're better off just freeing the memory directly
at the end of our collector pipeline rather than trying to
recycle it, especially if we have libtcmalloc available to us.

Also, set a RCVTIMEO option on the encoder worker pull sockets
so that they can have a limited amount of blocking on read. This
lets us avoid using zmq_poll, but still not use 100% CPU all the
time checking for incoming messages on each socket.