Don't waste time trying to maintain record alignment in export buf.
We're better off just trying to send the whole thing in one go,
as the entire buffer should already be record-aligned. If we only
send part of the buffer, we already have partialfront to help us
deal with recovering from subsequent disconnects etc.

The only real downside is that we are going to be buffering records
until we manage to successfully send the entire buffer contents,
even if most of the records in our "partialfront" section of the
buffer are no longer needed. Ideally, our send rate should be high
enough that partial sends happen very rarely, so our buffer
should be regularly drained and cleared. If not, the extra data
buffering is probably only one of many problems that we're going to
have, e.g. full queues, dropped packets, etc.