Separate collector config init+destroy from other state init+destroy
The main distinction is that state is only created and destroyed
once per collector instance -- once when the collector starts and
once when it finishes.

Config can be generated and destroyed multiple times, if the user
tells us to reload our configuration. We don't want to be creating
collector thread state variables or control queues when we just
want to grabbing a fresh config to compare against what we're
currently running.

Hence, I've separated init and destroy code that is related to
configuration (e.g. operator ID strings) into a different function
to the init and destroy code for message queues, thread local
state and other "operational state".