Fix problems where changing operatorID would have no effect.
Changing the operatorID, networkElementID or interceptionPointID
in the collector config and reloading it was not causing these
fields to change in subsequent exported ETSI records.

This was because we had pre-encoded these ETSI fields when the
intercept first began and did not make any attempt to update them
if the running config was to change.

These changes resolve the issue by adding a new publish message
type which is used to tell seqtracker threads that the global
intercept details may have changed. This will cause the tracker
thread to create a fresh set of preencoded ETSI fields for each
intercept that it knows about. The old preencoded fields are put
on the "removed" intercepts list for subsequent deletion (once
we're reasonably sure that there are no more encode jobs that
might be using a reference to the old preencoded fields).

Added a purge method that should tidy up the removed intercepts
list once a certain amount of time has passed. It is currently
run every 10K CC and/or IRI jobs that are created by the tracker
thread.