Add some instructions on how to use DPDK with OpenLI
Most of the instructions deal with building DPDK and binding
an interface to use the DPDK drivers, plus building a version
of libtrace that supports DPDK.

Ideally, packaging of DPDK (via dpdk-dev), libtrace-dpdk and
OpenLI itself will make most of the instructions here redundant.

However, until then it makes the most sense to give a full set
of explicit steps to follow to minimise the chances of stuff
being missed.

Note: this is a first draft, expect plenty of tweaks and changes
as we find stuff that I missed out or left too ambiguous.