Add config option to ignore RTP comfort noise packets.
At least one VOIP vendor produces comfort noise packets that is
treated as invalid by an agency decoder, so we need to allow users
to specify that comfort noise packets should not be intercepted
to avoid ruining the whole intercept.

To support this option, I've added an "options" bitmask to each
intercept which can be used to tell collectors which boolean
options are applicable to a given intercept. In addition, I've
added a mechanism for pushing modified values for the options
to the collectors so an option can be reconfigured on the fly
as needed without restarting any components.