Avoid repeated logging of handover failures
Specifically, if a handover no longer responds to keepalives for
some reason, we can end up filling our logs with endless disconnect
messages.

Now, we set a flag when a handover disconnects and avoid logging
any further messages for that handover until it successfully
gets a KA response from the agency.