Make sure we re-enable logging in cases where keepalives are disabled
If just keepalive responses are disabled / not expected, we
re-enable logging for a handover as soon as it sends a keepalive.

If keepalives are disabled, we re-enable logging as soon as the
handover re-connects. It might instantly fall over again, though,
so we might need to do a bit more in this case :/