Prevent log spam from provisioner if bad clients keep connecting
Same as with the mediator: clients that are dropped (either
mediators or collectors) have their IP address marked as bad
until such time as they are able to successfully auth with
the provisioner. We will not log any errors for a client with a
bad IP address so if something continually connects to us and
sends us garbage, we don't fill up our logs with error messages.