Reduce log spam in collector sync thread.
If the connection between the sync thread and provisioner breaks
down for some reason (i.e. the provisioner is not really a
provisioner or is sending us bogus data), we now only log an error
once and then suppress any subsequent errors until the provisioner
manages to send us something valid.