More log-spam reduction in the collector sync threads
Mostly this attempts to reduce log spam in the VOIP sync thread,
where we concern ourselves with two cases. The first is the case
where the provisioner is sending us bogus instructions -- this
is dealt with in the same way as earlier situations.

The second is the case where the SIP traffic that we are parsing
is not valid SIP (at least, according to our SIP parser). For now,
I've taken the approach that I log the first instance of invalid
SIP and then avoid logging any more unless the collector is HUPed.
However, all invalid SIP will still be logged to a trace file if
that is configured by the user.

Mainly I'm just worried about SIP implementations that do a lot
of weird stuff which will cause us to write a lot of
useless "invalid SIP" log messages -- the first log message
should be enough to tell the user that there is something weird in
their SIP traffic that they may want to get us to look into.