Log warnings every time the export buffer size grows another gigabyte
Previously we'd warn once we passed 1GB and never again.

Also removed error message when a transmit from an export buffer
fails -- the caller is better equipped to log the error,
especially in cases where the caller might want to suppress the
error message (due to the recipient having been unresponsive for
some time).