mediator: only log once for each "unknown" LIID that we see.
If a collector sends us an ETSI record with an LIID that is not
in our LIID->agency map, only log this mismatch once rather than
everytime we get a record for that LIID.

Also in this case, do NOT drop the connection to the collector.
Instead, just discard the record.