Allow gitlab packaging to fail for sid and buster
Sid and buster are pretty cutting edge and prone to flakiness
as a result -- we've already had problems with the libtrace
package for buster/sid due to dpdk dependencies that have
disappeared since the last libtrace release. Since most users
are going to be installing on something a bit more stable
(hopefully), we don't want to fail the whole packaging process
just because of that.