Fix minor memory leaks when duplicate configuration occurs
Also rework all string duplication of configuration values to be
consistent across all config options. Specifically, if an option
appears multiple times when it is only expected once, we will
always use the lattermost instance of the configuration option
and make sure to free any earlier strdup() that we have performed
for that config.