Collector: don't panic if the target identifier for an intercept changes
Fixes bug where changing the username or ALU shim id for an IP
intercept, while still using the same LIID, will cause collectors
to continually disconnect and reconnect to the provisioner.

While it was a good idea to try and enforce uniqueness of LIIDs
after intercept configuration changes, this didn't account for
the possibility of human error during intercept configuration.
A person can easily typo a username and will want to go back and
correct it -- in that case, they will probably still need to use
the same LIID for the intercept. Our previous code would not allow
that and in fact would fail horribly if the user tried to do so.

The new approach is simply to warn that the LIID is being reused
in the log messages and accept the new username or shim ID as
being the correct one.