Fix bug preventing IPv6 packets from being intercepted
For some reason, when we moved to the packet_info_t method of
extracting and storing IP addresses and ports from captured
packets, we forgot to update the IPv6 code path properly. As a
result, what we thought were the IPv6 addresses for the current
packet were actually uninitialised memory.