Add support for decapsulating JMirror packets
JMirror is an encapsulation similar to the ALU shim that we
already support, which is used by Juniper devices to stream
"LI" intercepted packets to a device for encoding and
mediation.

Most of the code here is a copy-paste-modify of the ALU shim
decapsulation code, as the formats are very similar.

I've moved some common code to places where it is more accessible
to both decapsulators. There is definitely more scope to reduce
code replication between the two, and I may decide to do that
later on, but for now we have at least a working implementation
(tested against a single example trace).