Re-factor ALU and JMirror code to remove lots of duplication
These two code paths are very similar. Now, instead of having a
distinction between ALU intercepts and JMirror intercepts, we
have a single "vendor mirror" intercept type that covers both.

I've still kept the list of ALU mirrors and Juniper mirrors
separate -- the shim headers are interpreted slightly differently
for each vendor, so we need to be able to recognise when we've
matched one over the other.

Note: there is one potential situation where my changes could
go awry. If a collector is receiving BOTH ALU and Jmirror traffic
and both streams use the same intercept ID (but for different
customers) then anything with that intercept ID is probably going
to be treated as the ALU intercept (even if it comes from the
Juniper mirror).

However, this is extremely unlikely -- who's mixing and matching
vendors like that? -- and can be easily resolved by simply NOT
using the same intercept ID for multiple intercepts across the
two vendors.