Provisioner: separate intercept/agency config from provisioner config
In the near future, intercept config is going to be managed
entirely by the provisioner -- i.e. manual edits will be
strongly discouraged. Any changes to intercept config will
be written to the appropriate config file by the provisioner
itself (or via a stand-alone CLI app).

However, we still want to allow users to tweak some of the
provisioner configuration (e.g. ports for communication with
other components etc), so these options will stay in the original
config file which will otherwise behave the same as it previously
did (i.e. be reloadable via HUP).