openlicollector: only send "intercept details" for new VOIP intercepts
Previously, we were always sending a details message for each VOIP
intercept that was announced to us, even if we had already seen
that LIID before.

This was causing the forwarder thread to reset
its expected sequence number for that LIID, but the rest of the
collection system was continuing from the old sequence number.
If we'd intercepted at least one packet for that target, the
two sequence numbers would get out of sync.

This, in turn, meant that no more records would ever be sent to
the mediator for that intercept as none of the incoming records
would ever have the expected sequence number of 0.