provisioner: "username" is now compulsory for all IP intercepts
This solves a problem where ALU / Jmirror users would forget
to set a user and wonder why they were getting no useful output
from HI2.

This does mean that we are enforcing user names for static IP
intercepts, which may not quite make sense but presumably there
is some useful value that operators can put in here in those
cases anyway (as long as it is not a name that belongs to another
user that is using AAA).

This should also help with the problem where a username is
added to an intercept config that was missing one, but the
provisioner would not recognise that the intercept has changed and
therefore not push that new information down to the collectors.